test(platform): fix zero-directed-connect-page tests per review
- Remove implementation-detail tests (CONN-D-043, CONN-C-046) that assert on .animate-spin CSS class
- Remove over-testing case (CONN-I-048) that only verifies React controlled input behavior
- Move valid new tests (CONN-D-045, CONN-I-047, CONN-I-049) into the canonical test file at views/zero-page/__tests__/
- Delete misplaced test file from views/conn/__tests__/
